Made calling util.optimizer more save against spelling errors.
Previously, if an optimiser was created with this function and an argument in kwargs was supplied which was not expected by the optimiser klass, this passed silently. This is dangerous wrt to arguments like 'steprate' which might be mistyped as 'step_rate' and thus be ignored. From now on, each unused argument issues a warning.