Corrected documentation
Square roots in the Nesterov momentum was missing.
minusses added in front of the velocity factors, can also be changed in the theta update. Otherwise the algorithm walks away from the optimum.
Documentation changes proposed because the readthedocs website is a great source for the code for these optimization algorithms!